fix(rust): satisfy introduced-change policy
